requirements.txt
openai
python-dotenv
requests
pillow
opencv-python
flask
